Start of Vendor Program.
Some small changes to various parts of the program. May be difference
in how I and Guilherme did Credit Cards, so that will be changed as
needed. The purchaseSession() function in Vendor.cpp has the meat of
the program. It is static so it can function as a ‘stand-alone’ in any
program.