Wipes Transactions at end of month
Desired feature. Currently, the balance on the card is NOT reset as the
limit is not implemented anyways. Eventually, balance would need to be
separated from monthly availability, but I haven’t done this yet. It
also thus displays the balance on the card even when there’s no
transactions. This is required because someone could have a minimum
purchase history, so the balance is not 0 after the update.